https://github.com/manifold-systems/manifold/issues/35
The essence of this issue is about handling a *call site* where the receiver is a type variable bounded by a structural interface.  The change enables such a call site to work.

At a higher level the remainder of the issue is about supporting assignability wrt structural variance at the type variable bounds.  This is already supported by adding the structural interface to the type via extension.  More detailed info follows in https://github.com/manifold-systems/manifold/issues/35